+++ packages/ui-components/src/components/Maps/USStateMap/MetricUSStateMap.tsx
@@ -2,16 +2,16 @@ import React from "react";
 import { USStateMap } from "./USStateMap";
 import { MetricUSStateMapProps } from "./interfaces";
 import { useDataForRegionsAndMetrics } from "../../../common/hooks";
-import { states } from "@actnowcoalition/regions";
 import { getCountiesOfState } from "../../../common/utils/maps";
 
 export const MetricUSStateMap: React.FC<MetricUSStateMapProps> = ({
   metric,
   stateRegionId,
+  regionDB,
   ...otherProps
 }) => {
-  const state = states.findByRegionIdStrict(stateRegionId);
-  const countiesOfState = getCountiesOfState(stateRegionId);
+  const state = regionDB.findByRegionIdStrict(stateRegionId);
+  const countiesOfState = getCountiesOfState(regionDB, stateRegionId);
   const mapRegions = [...countiesOfState, state];
 
   const { data } = useDataForRegionsAndMetrics(mapRegions, [metric], false);
@@ -22,7 +22,10 @@ export const MetricUSStateMap: React.FC<MetricUSStateMapProps> = ({
 
   return (
     <USStateMap
-      getFillColor={(region) => data.metricData(region, metric).getColor()}
+      getFillColor={(regionId: string) => {
+        const region = regionDB.findByRegionId(regionId);
+        return region ? data.metricData(region, metric).getColor() : "#eee";
+      }}
       stateRegionId={stateRegionId}
       {...otherProps}
     />