+++ packages/ui-components/src/components/SocialMetaTags/index.ts
@@ -1,3 +1,0 @@
-export * from "./PageMetaTags";
-export * from "./FacebookMetaTags";
-export * from "./TwitterMetaTags";