+++ packages/metrics/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @actnowcoalition/metrics
 
+## 0.2.0
+
+### Minor Changes
+
+- 7d4bd88: Change the default for includeTimeseries to false when fetching data and fix the MetricCompareTable, MetricDot, and MetricValue to not fetch timeseries.
+
+## 0.1.1
+
+### Patch Changes
+
+- a342e24: Add node-fetch dependency so that metrics can be fetched from node.js apps.
+- 464f4a4: Avoid re-fetching CSV files when a fetch is already in progress.
+- Updated dependencies [e8f6d2b]
+  - @actnowcoalition/regions@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes