+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @actnowcoalition/ui-components
 
+## 0.3.0
+
+### Minor Changes
+
+- 7d4bd88: Change the default for includeTimeseries to false when fetching data and fix the MetricCompareTable, MetricDot, and MetricValue to not fetch timeseries.
+
+### Patch Changes
+
+- Updated dependencies [7d4bd88]
+  - @actnowcoalition/metrics@0.2.0
+
+## 0.2.3
+
+### Patch Changes
+
+- 05fa5e7: Implement tabs
+- Updated dependencies [e8f6d2b]
+- Updated dependencies [a342e24]
+- Updated dependencies [464f4a4]
+  - @actnowcoalition/regions@0.1.1
+  - @actnowcoalition/metrics@0.1.1
+
 ## 0.2.2
 
 ### Patch Changes